Move cli module
