+++ scripts/3.2.1_assign_primary_zone_edu.py
@@ -2,11 +2,11 @@ import geopandas as gpd
 import pandas as pd
 
 import acbm
-from acbm.assigning.cli import acbm_cli
 from acbm.assigning.select_zone_primary import (
     fill_missing_zones,
     select_zone,
 )
+from acbm.cli import acbm_cli
 from acbm.logger_config import assigning_primary_zones_logger as logger
 from acbm.preprocessing import add_location
 from acbm.utils import get_config, init_rng
