+++ scripts/3.2.2_assign_primary_zone_work.py
@@ -4,13 +4,13 @@ import geopandas as gpd
 import pandas as pd
 
 import acbm
-from acbm.assigning.cli import acbm_cli
 from acbm.assigning.plots import (
     plot_workzone_assignment_heatmap,
     plot_workzone_assignment_line,
 )
 from acbm.assigning.select_zone_work import WorkZoneAssignment
 from acbm.assigning.utils import filter_matrix_to_boundary
+from acbm.cli import acbm_cli
 from acbm.logger_config import assigning_primary_zones_logger as logger
 from acbm.preprocessing import add_locations_to_activity_chains
 from acbm.utils import calculate_rmse, get_config, init_rng
